pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "pfille", email = "pfille@posteo.de"},
]
dependencies = [
    "py3langid>=0.2.1",
    "loguru>=0.6.0",
    "pillow>=9.1.1",
    "PyHyphen>=4.0.3",
    "socketIO-client>=0.7.2",
    "PyYAML>=6.0",
    "python-socketio[client]==4.6.1",
]
requires-python = ">=3.9"

[tool.pdm]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
